+++ src/Meta_Model.h
@@ -205,7 +205,7 @@ class META_MODEL_DECL_SPEC Meta_Model : public class_base
 
    void clear( );
 
-   void validate( unsigned state, bool is_internal );
+   void validate( uint64_t state, bool is_internal );
    void validate_set_fields( std::set< std::string >& fields_set );
 
    void after_fetch( );
