+++ src/Meta_Module.cpp
@@ -435,7 +435,7 @@ struct Meta_Module::impl : public Meta_Module_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -804,7 +804,8 @@ bool Meta_Module::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Module::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Module::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -831,7 +832,8 @@ void Meta_Module::impl::validate( unsigned state, bool is_internal, validation_e
    // [<finish validate>]
 }
 
-void Meta_Module::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Module::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -1169,7 +1171,7 @@ void Meta_Module::clear( )
    p_impl->clear( );
 }
 
-void Meta_Module::validate( unsigned state, bool is_internal )
+void Meta_Module::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
