+++ src/Meta_Procedure.cpp
@@ -741,7 +741,7 @@ struct Meta_Procedure::impl : public Meta_Procedure_command_handler
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -1079,7 +1079,8 @@ bool Meta_Procedure::impl::value_will_be_provided( const string& field_name )
    return false;
 }
 
-void Meta_Procedure::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Procedure::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -1126,7 +1127,8 @@ void Meta_Procedure::impl::validate( unsigned state, bool is_internal, validatio
    // [<finish validate>]
 }
 
-void Meta_Procedure::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Procedure::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -1713,7 +1715,7 @@ void Meta_Procedure::clear( )
    p_impl->clear( );
 }
 
-void Meta_Procedure::validate( unsigned state, bool is_internal )
+void Meta_Procedure::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
