+++ src/Meta_Specification_Type.cpp
@@ -3370,7 +3370,7 @@ struct Meta_Specification_Type::impl : public Meta_Specification_Type_command_ha
 
    bool value_will_be_provided( const string& field_name );
 
-   void validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors );
+   void validate( uint64_t state, bool is_internal, validation_error_container* p_validation_errors );
    void validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors );
 
    void after_fetch( );
@@ -5978,7 +5978,8 @@ bool Meta_Specification_Type::impl::value_will_be_provided( const string& field_
    return false;
 }
 
-void Meta_Specification_Type::impl::validate( unsigned state, bool is_internal, validation_error_container* p_validation_errors )
+void Meta_Specification_Type::impl::validate(
+ uint64_t state, bool is_internal, validation_error_container* p_validation_errors )
 {
    ( void )state;
    ( void )is_internal;
@@ -6071,7 +6072,8 @@ void Meta_Specification_Type::impl::validate( unsigned state, bool is_internal, 
    // [<finish validate>]
 }
 
-void Meta_Specification_Type::impl::validate_set_fields( set< string >& fields_set, validation_error_container* p_validation_errors )
+void Meta_Specification_Type::impl::validate_set_fields(
+ set< string >& fields_set, validation_error_container* p_validation_errors )
 {
    ( void )fields_set;
 
@@ -7673,7 +7675,7 @@ void Meta_Specification_Type::clear( )
    p_impl->clear( );
 }
 
-void Meta_Specification_Type::validate( unsigned state, bool is_internal )
+void Meta_Specification_Type::validate( uint64_t state, bool is_internal )
 {
    p_impl->validate( state, is_internal, &validation_errors );
 }
